Update all projects for Xcode 8 + Swift 3.

+++ Lesson 38/Lesson 38/AppController.m
@@ -3,7 +3,7 @@
 //  Lesson 38
 //
 //  Created by Lucas Derraugh on 4/29/12.
-//  Copyright (c) 2012 __MyCompanyName__. All rights reserved.
+//  Copyright (c) 2012 Lucas Derraugh. All rights reserved.
 //
 
 #import "AppController.h"
@@ -11,7 +11,7 @@
 @implementation AppController
 
 - (IBAction)activateAlert:(id)sender {
-    int num = 6;
+    NSInteger num = 6;
     [[NSAlert alertWithMessageText:NSLocalizedString(@"Message", @"This is a message")
                     defaultButton:NSLocalizedString(@"Yes", @"Yes") 
                   alternateButton:NSLocalizedString(@"No", @"No")
